+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## 0.0.10 (unreleased)
-- Add ability to hide items if they cannot be achived in the current save due to missing prerequisites - work in progress
+- Add ability to display in italic items that cannot be obtained in the current save, because they depend on something else not avialable in the current save that the character does not currently have
 
 ## 0.0.9 (10 May 2023)
 - Small fix for adventure panel not updating